[ALPS09196978] merge back mainline changes
[Description]
clone a new kernel_device_modules-mainline from 6.6
and merge back the changes of current kernel_device_modules-mainline

clone from: 14eccbba6291 of kernel_device_modules-6.6
(alps-mp-v0.mp1.cs1--2024_08_23_00_00)
merge: 5098bea9f9db of kernel_device_modules-mainline
(alps-mp-v0.mp1--2024_08_20_00_00)

patchset 1: merge with conflicts
patchset 2: resolve conflicts
patchset 3: switch to mainline
patchset 4: add back gzvm driver
patchset 5: remove lockdep_off()
patchset 6: configs and bzl

(fixing build errors)
patchset 7: linux/of headers
patchset 8: strlcpy => strscpy
patchset 9: asoc_* => snd_soc_*
patchset 10: SWIP-00000012 CPU QoS Driver
patchset 11: SWIP-00000062 iommu_kernel
patchset 12: SWIP-00000078 logstore
patchset 13: SWIP-00000099 ethernet_dwmac
patchset 14: SWIP-00000121 Energy Aware Scheduling Plus (EAS+)
patchset 15: SWIP-00000146 ufs
patchset 16: SWIP-00000147 usb_ssusb
patchset 17: SWIP-00000151 Video Driver
patchset 18: SWIP-00000195 thermal driver
patchset 19: SWIP-00000216 dma_buf_kernel
patchset 20+: other minor changes

[Test]
build pass

MTK-Commit-Id: 50fa483e786ac9b1a43b45166e87f81c05e74587

CR-Id: ALPS09196978
Feature: Kernel SI Operation

+++ drivers/misc/mediatek/cmdq/mailbox/cmdq-sec-mailbox.c
@@ -14,6 +14,7 @@
 #include <linux/sched/clock.h>
 #include <linux/timer.h>
 #include <linux/of_device.h>
+#include <linux/of_platform.h>
 
 #include "cmdq-sec.h"
 #include "cmdq-sec-mailbox.h"
@@ -252,7 +253,7 @@ static inline void cmdq_mmp_init(struct cmdq_sec *cmdq)
 
 	len = snprintf(name, sizeof(name), "cmdq_sec_%hhu", cmdq->hwid);
 	if (len >= sizeof(name))
-		cmdq_log("len:%d over name size:%lu", len, (unsigned long)(sizeof(name)));
+		cmdq_log("len:%d over name size:%lu", len, sizeof(name));
 
 	cmdq->mmp.cmdq_root = mmprofile_register_event(MMP_ROOT_EVENT, "CMDQ");
 	cmdq->mmp.cmdq = mmprofile_register_event(cmdq->mmp.cmdq_root, name);
@@ -1751,7 +1752,7 @@ static int cmdq_sec_mbox_startup(struct mbox_chan *chan)
 	INIT_WORK(&thread->timeout_work, cmdq_sec_task_timeout_work);
 	len = snprintf(name, sizeof(name), "task_exec_wq_%u", thread->idx);
 	if (len >= sizeof(name))
-		cmdq_log("len:%d over name size:%lu", len, (unsigned long)(sizeof(name)));
+		cmdq_log("len:%d over name size:%lu", len, sizeof(name));
 
 	thread->task_exec_wq = create_singlethread_workqueue(name);
 	thread->occupied = true;
